Refactor generic code into a shared Agent base class
